feat(tauri): built-in bundled self-hosted-node
