+++ src/platform/tauri.ts
@@ -106,8 +106,8 @@ class TauriServerApi implements IServerApi {
     return invoke<void>("server_list_players");
   }
 
-  async createRoom(params: CreateRoomParams): Promise<void> {
-    return invoke<void>("server_create_room", {
+  async createRoom(params: CreateRoomParams): Promise<string | null> {
+    return invoke<string | null>("server_create_room", {
       roomName: params.roomName,
       maxPlayers: params.maxPlayers,
       format: params.format,
@@ -116,7 +116,12 @@ class TauriServerApi implements IServerApi {
       draftConfig: params.draftConfig ?? null,
       sealedConfig: params.sealedConfig ?? null,
       reconnectTimeoutS: params.reconnectTimeoutS ?? null,
+      password: params.password ?? null,
     });
+  }
+
+  async stopRoom(): Promise<void> {
+    return invoke<void>("server_stop_room");
   }
 
   async joinRoom(params: JoinRoomParams): Promise<void> {
