+++ src/views/Lobby.tsx
@@ -107,6 +107,7 @@ export default function Lobby() {
     playerId,
     rooms,
     currentRoom,
+    roomPassword,
     players,
     gameStarted,
     playerOrder,
@@ -573,6 +574,7 @@ export default function Lobby() {
           <TablesList
             rooms={rooms}
             currentRoom={currentRoom}
+            roomPassword={roomPassword}
             username={username}
             onNewGame={() => setCreateRoomOpen(true)}
             onRefresh={refreshLobbyData}
